Consolidate address balance and faucet endpoints
This consolidates the address balance and faucet endpoints to
use the non-wallet scoped ones.
