+++ spec/spec_helper.rb
@@ -11,4 +11,5 @@ SimpleCov.start do
 end
 
 require_relative '../lib/coinbase'
+require_relative 'support/shared_examples/address_balances'
 require_relative 'support/shared_examples/pagination'
