+++ spec/unit/coinbase/address/wallet_address_spec.rb
@@ -24,13 +24,13 @@ describe Coinbase::WalletAddress do
   let(:weth_asset) do
     Coinbase::Client::Asset.new(network_id: normalized_network_id, asset_id: 'weth', decimals: 18)
   end
-  let(:addresses_api) { double('Coinbase::Client::AddressesApi') }
+  let(:addresses_api) { double('Coinbase::Client::ExternalAddressesApi') }
   let(:assets_api) { double('Coinbase::Client::AssetsApi') }
   let(:transfers_api) { double('Coinbase::Client::TransfersApi') }
   let(:trades_api) { double('Coinbase::Client::TradesApi') }
 
   before(:each) do
-    allow(Coinbase::Client::AddressesApi).to receive(:new).and_return(addresses_api)
+    allow(Coinbase::Client::ExternalAddressesApi).to receive(:new).and_return(addresses_api)
     allow(Coinbase::Client::AssetsApi).to receive(:new).and_return(assets_api)
     allow(Coinbase::Client::TransfersApi).to receive(:new).and_return(transfers_api)
     allow(Coinbase::Client::TradesApi).to receive(:new).and_return(trades_api)
@@ -113,96 +113,9 @@ describe Coinbase::WalletAddress do
     end
   end
 
-  describe '#balances' do
-    let(:response) do
-      Coinbase::Client::AddressBalanceList.new(
-        data: [
-          Coinbase::Client::Balance.new(amount: '1000000000000000000', asset: eth_asset),
-          Coinbase::Client::Balance.new(amount: '5000000000', asset: usdc_asset),
-          Coinbase::Client::Balance.new(amount: '3000000000000000000', asset: weth_asset)
-        ]
-      )
-    end
+  it_behaves_like 'an address that supports balance queries'
 
-    it 'returns a hash with balances' do
-      expect(addresses_api)
-        .to receive(:list_address_balances)
-        .with(wallet_id, address_id)
-        .and_return(response)
-
-      expect(address.balances).to eq(
-        eth: BigDecimal('1'),
-        usdc: BigDecimal('5000'),
-        weth: BigDecimal('3')
-      )
-    end
-  end
-
-  describe '#balance' do
-    let(:response) do
-      Coinbase::Client::Balance.new(amount: '1000000000000000000', asset: eth_asset)
-    end
-
-    before do
-      allow(addresses_api)
-        .to receive(:get_address_balance)
-        .with(wallet_id, address_id, primary_denomination)
-        .and_return(response)
-    end
-
-    context 'when the asset_id is :eth' do
-      let(:asset_id) { :eth }
-      let(:primary_denomination) { 'eth' }
-
-      it 'returns the correct ETH balance' do
-        expect(address.balance(:eth)).to eq BigDecimal('1')
-      end
-    end
-
-    context 'when the asset_id is :gwei' do
-      let(:asset_id) { :gwei }
-      let(:primary_denomination) { 'eth' }
-
-      it 'returns the correct Gwei balance' do
-        expect(address.balance(:gwei)).to eq BigDecimal('1_000_000_000')
-      end
-    end
-
-    context 'when the asset_id is :wei' do
-      let(:asset_id) { :wei }
-      let(:primary_denomination) { 'eth' }
-
-      it 'returns the correct Wei balance' do
-        expect(address.balance(:wei)).to eq BigDecimal('1_000_000_000_000_000_000')
-      end
-    end
-
-    context 'when the asset id is a non-eth denomination' do
-      let(:asset_id) { :other }
-      let(:primary_denomination) { 'other' }
-      let(:decimals) { 7 }
-      let(:other_asset) do
-        Coinbase::Client::Asset.new(network_id: 'base-sepolia', asset_id: 'other', decimals: decimals)
-      end
-      let(:response) do
-        Coinbase::Client::Balance.new(amount: '1000000000000000000', asset: other_asset)
-      end
-
-      it 'returns the correct balance' do
-        expect(address.balance(:other)).to eq BigDecimal('100_000_000_000')
-      end
-    end
-
-    context 'when there is no response' do
-      let(:response) { nil }
-      let(:asset_id) { :eth }
-      let(:primary_denomination) { 'eth' }
-
-      it 'returns 0' do
-        expect(address.balance(:eth)).to eq BigDecimal('0')
-      end
-    end
-  end
+  it_behaves_like 'an address that supports requesting faucet funds'
 
   describe '#transfer' do
     let(:eth_balance_response) do
@@ -277,8 +190,8 @@ describe Coinbase::WalletAddress do
 
       before do
         allow(addresses_api)
-          .to receive(:get_address_balance)
-          .with(wallet_id, address_id, transfer_asset_id)
+          .to receive(:get_external_address_balance)
+          .with(normalized_network_id, address_id, transfer_asset_id)
           .and_return(balance_response)
 
         allow(transfers_api)
@@ -466,8 +379,8 @@ describe Coinbase::WalletAddress do
 
       before do
         expect(addresses_api)
-          .to receive(:get_address_balance)
-          .with(wallet_id, address_id, 'eth')
+          .to receive(:get_external_address_balance)
+          .with(normalized_network_id, address_id, 'eth')
           .and_return(eth_balance_response)
       end
 
@@ -506,8 +419,8 @@ describe Coinbase::WalletAddress do
         allow(Coinbase).to receive(:configuration).and_return(configuration)
 
         allow(addresses_api)
-          .to receive(:get_address_balance)
-          .with(wallet_id, address_id, transfer_asset_id)
+          .to receive(:get_external_address_balance)
+          .with(normalized_network_id, address_id, transfer_asset_id)
           .and_return(balance_response)
 
         allow(transfers_api)
@@ -519,8 +432,8 @@ describe Coinbase::WalletAddress do
 
       it 'creates a transfer without broadcast' do
         expect(addresses_api)
-          .to receive(:get_address_balance)
-          .with(wallet_id, address_id, 'eth')
+          .to receive(:get_external_address_balance)
+          .with(normalized_network_id, address_id, 'eth')
           .and_return(eth_balance_response)
         expect(transfers_api)
           .to receive(:create_transfer)
@@ -621,8 +534,8 @@ describe Coinbase::WalletAddress do
 
       before do
         allow(addresses_api)
-          .to receive(:get_address_balance)
-          .with(wallet_id, address_id, normalized_from_asset_id)
+          .to receive(:get_external_address_balance)
+          .with(normalized_network_id, address_id, normalized_from_asset_id)
           .and_return(balance_response)
 
         allow(trades_api)
@@ -743,8 +656,8 @@ describe Coinbase::WalletAddress do
 
       before do
         expect(addresses_api)
-          .to receive(:get_address_balance)
-          .with(wallet_id, address_id, 'eth')
+          .to receive(:get_external_address_balance)
+          .with(normalized_network_id, address_id, 'eth')
           .and_return(eth_balance_response)
       end
 
@@ -755,71 +668,6 @@ describe Coinbase::WalletAddress do
           ArgumentError,
           "Insufficient funds: #{excessive_amount} requested, but only #{current_balance} available"
         )
-      end
-    end
-  end
-
-  describe '#faucet' do
-    let(:request) { double('Request', transaction: transaction) }
-    let(:tx_hash) { '0xdeadbeef' }
-    let(:faucet_tx) do
-      instance_double('Coinbase::Client::FaucetTransaction', transaction_hash: tx_hash)
-    end
-
-    context 'when the request is successful' do
-      subject(:faucet_response) { address.faucet }
-
-      before do
-        expect(addresses_api)
-          .to receive(:request_faucet_funds)
-          .with(wallet_id, address_id)
-          .and_return(faucet_tx)
-      end
-
-      it 'requests funds from the faucet and returns the faucet transaction' do
-        expect(faucet_response).to be_a(Coinbase::FaucetTransaction)
-        expect(faucet_response.transaction_hash).to eq(tx_hash)
-      end
-    end
-
-    context 'when the request is unsuccesful' do
-      before do
-        expect(addresses_api)
-          .to receive(:request_faucet_funds)
-          .with(wallet_id, address_id)
-          .and_raise(api_error)
-      end
-
-      context 'when the faucet limit is reached' do
-        let(:api_error) do
-          Coinbase::Client::ApiError.new(
-            code: 429,
-            response_body: {
-              'code' => 'faucet_limit_reached',
-              'message' => 'failed to claim funds - address likely has already claimed in the past 24 hours'
-            }.to_json
-          )
-        end
-
-        it 'raises a FaucetLimitReachedError' do
-          expect { address.faucet }.to raise_error(::Coinbase::FaucetLimitReachedError)
-        end
-      end
-
-      context 'when the request fails unexpectedly' do
-        let(:api_error) do
-          Coinbase::Client::ApiError.new(
-            code: 500,
-            response_body: {
-              'code' => 'internal',
-              'message' => 'unexpected error occurred while requesting faucet funds'
-            }.to_json
-          )
-        end
-
-        it 'raises an internal error' do
-          expect { address.faucet }.to raise_error(::Coinbase::InternalError)
-        end
       end
     end
   end
